refactor: Refactored DeleteTeamMember usecase

+++ apps/api/src/app/project-invitation/usecase/delete-team-member/delete-team-member.usecase.ts
@@ -1,11 +1,19 @@
-import { EnvironmentRepository } from '@impler/dal';
 import { Injectable } from '@nestjs/common';
+import { EnvironmentRepository, UserRepository } from '@impler/dal';
+import { DocumentNotFoundException } from '@shared/exceptions/document-not-found.exception';
 
 @Injectable()
 export class DeleteTeamMember {
-  constructor(private environmentRepository: EnvironmentRepository) {}
+  constructor(
+    private environmentRepository: EnvironmentRepository,
+    private userRepository: UserRepository
+  ) {}
 
   async exec({ projectId, userId }: { projectId: string; userId: string }) {
-    return await this.environmentRepository.deleteTeamMember(projectId, userId);
+    const teamMember = await this.userRepository.findById(userId, 'firstName lastName email profilePicture');
+    if (!teamMember) throw new DocumentNotFoundException('TeamMember', userId);
+    await this.environmentRepository.deleteTeamMember(projectId, userId);
+
+    return teamMember;
   }
 }
